Abstract gulp paths and prepare for reorganisation
